Extract types to file
