+++ lib/wrappers/gltf-renderer/primitive.types.ts
@@ -1,0 +1,10 @@
+import type { GlTFMaterial } from '../../gltf/material.types';
+import type { Primitive } from '../../gl/primitive';
+import type { Mat4 } from '../../geometry/mat4.types';
+
+export interface PrimitiveWrapper {
+    readonly primitive: Primitive;
+    readonly matrix: Mat4;
+    readonly normalMatrix: Mat4;
+    readonly material: GlTFMaterial | null;
+}
